requirements.txt
hikari-lightbulb
pycoingecko